requirements.txt
black
mypy
tqdm
rich
mbforbes_python_utils
requests
types-requests
python-frontmatter
mistune
beautifulsoup4
